+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Fix: Add props `others` on `LinkList` component.
+- Fix: Add `others` props on `LinkList` component.
 
 ## [21.0.0] - 2018-06-27